fix : refactor base.dart to ui directory
